chore: reorganize routes for consistency
