+++ webapp/app/routes/_index/_route.tsx
@@ -63,7 +63,7 @@ export default function Index() {
                 />
               ) : (
                 <div className="flex space-x-4">
-                  <Link to={$path("/app")}>
+                  <Link to={$path("/app/proposals")}>
                     <Button>Standard Upgrades</Button>
                   </Link>
                   <Link to={$path("/app/emergency")}>
