+++ webapp/app/routes/app/proposals/$id/_route.tsx
@@ -8,6 +8,7 @@ import { validateAndSaveProposalSignature } from "@/.server/service/signatures";
 import { hexSchema } from "@/common/basic-schemas";
 import { signActionSchema } from "@/common/sign-action";
 import TxLink from "@/components/tx-link";
+import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import Loading from "@/components/ui/loading";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
@@ -22,6 +23,7 @@ import { RawStandardUpgrade } from "@/routes/app/proposals/$id/raw-standard-upgr
 import SignButton from "@/routes/app/proposals/$id/sign-button";
 import SystemContractChangesTable from "@/routes/app/proposals/$id/system-contract-changes-table";
 import { requireUserFromHeader } from "@/utils/auth-headers";
+import { displayBytes32 } from "@/utils/bytes32";
 import { compareHexValues } from "@/utils/compare-hex-values";
 import { dateToUnixTimestamp } from "@/utils/date";
 import { badRequest, notFound } from "@/utils/http";
@@ -29,7 +31,8 @@ import { PROPOSAL_STATES } from "@/utils/proposal-states";
 import { env } from "@config/env.server";
 import type { ActionFunctionArgs, LoaderFunctionArgs } from "@remix-run/node";
 import { defer, json } from "@remix-run/node";
-import { Await, useLoaderData } from "@remix-run/react";
+import { Await, useLoaderData, useNavigate } from "@remix-run/react";
+import { ArrowLeft } from "lucide-react";
 import { Suspense } from "react";
 import { getFormData, getParams } from "remix-params-helper";
 import { $path } from "remix-routes";
@@ -152,302 +155,318 @@ const NECESSARY_LEGAL_VETO_SIGNATURES = 2;
 
 export default function Proposals() {
   const { user, asyncData, proposalId } = useLoaderData<typeof loader>();
+  const navigate = useNavigate();
 
   return (
-    <div className="flex flex-1 flex-col space-y-4">
-      <Suspense
-        fallback={
-          <div className="flex flex-1 flex-col items-center justify-center space-y-6">
-            <Loading />
-            <h2>Loading proposal...</h2>
-          </div>
-        }
-      >
-        <Await resolve={asyncData}>
-          {({
-            addresses,
-            proposal,
-            reports,
-            userSignedLegalVeto,
-            userSignedProposal,
-            ethNetwork,
-          }) => {
-            const securityCouncilSignaturesReached =
-              proposal.signatures.approveUpgradeSecurityCouncil.length >=
-              NECESSARY_SECURITY_COUNCIL_SIGNATURES;
-            const guardiansSignaturesReached =
-              proposal.signatures.approveUpgradeGuardians.length >= NECESSARY_GUARDIAN_SIGNATURES;
+    <div className="mt-10 flex flex-1 flex-col">
+      <div className="mb-4 flex items-center pl-2">
+        <Button
+          size="icon"
+          variant="ghost"
+          onClick={() => navigate(-1)}
+          className="mr-2 hover:bg-transparent"
+        >
+          <ArrowLeft />
+        </Button>
+        <h2 className="font-semibold">Proposal {displayBytes32(proposalId)}</h2>
+      </div>
+      <div className="flex flex-1 flex-col space-y-4">
+        <Suspense
+          fallback={
+            <div className="flex flex-1 flex-col items-center justify-center space-y-6">
+              <Loading />
+              <h2>Loading proposal...</h2>
+            </div>
+          }
+        >
+          <Await resolve={asyncData}>
+            {({
+              addresses,
+              proposal,
+              reports,
+              userSignedLegalVeto,
+              userSignedProposal,
+              ethNetwork,
+            }) => {
+              const securityCouncilSignaturesReached =
+                proposal.signatures.approveUpgradeSecurityCouncil.length >=
+                NECESSARY_SECURITY_COUNCIL_SIGNATURES;
+              const guardiansSignaturesReached =
+                proposal.signatures.approveUpgradeGuardians.length >= NECESSARY_GUARDIAN_SIGNATURES;
 
-            const signProposalEnabled =
-              !userSignedProposal &&
-              [PROPOSAL_STATES.LegalVetoPeriod, PROPOSAL_STATES.Waiting].includes(
-                proposal.status
-              ) &&
-              (user.role === "guardian" ? !proposal.guardiansApproval : true);
-            const signLegalVetoEnabled =
-              !userSignedLegalVeto &&
-              !proposal.guardiansExtendedLegalVeto &&
-              proposal.status === PROPOSAL_STATES.LegalVetoPeriod &&
-              user.role === "guardian";
+              const signProposalEnabled =
+                !userSignedProposal &&
+                [PROPOSAL_STATES.LegalVetoPeriod, PROPOSAL_STATES.Waiting].includes(
+                  proposal.status
+                ) &&
+                (user.role === "guardian" ? !proposal.guardiansApproval : true);
+              const signLegalVetoEnabled =
+                !userSignedLegalVeto &&
+                !proposal.guardiansExtendedLegalVeto &&
+                proposal.status === PROPOSAL_STATES.LegalVetoPeriod &&
+                user.role === "guardian";
 
-            const executeSecurityCouncilApprovalEnabled =
-              proposal.status === PROPOSAL_STATES.Waiting && securityCouncilSignaturesReached;
-            const executeGuardiansApprovalEnabled =
-              !proposal.guardiansApproval &&
-              proposal.status === PROPOSAL_STATES.Waiting &&
-              guardiansSignaturesReached;
-            const executeLegalVetoExtensionEnabled =
-              !proposal.extendedLegalVeto && proposal.status === PROPOSAL_STATES.LegalVetoPeriod;
-            const executeProposalEnabled =
-              proposal.status === PROPOSAL_STATES.Ready &&
-              (isAddressEqual(proposal.executor as Hex, user.address as Hex) ||
-                isAddressEqual(proposal.executor as Hex, zeroAddress));
+              const executeSecurityCouncilApprovalEnabled =
+                proposal.status === PROPOSAL_STATES.Waiting && securityCouncilSignaturesReached;
+              const executeGuardiansApprovalEnabled =
+                !proposal.guardiansApproval &&
+                proposal.status === PROPOSAL_STATES.Waiting &&
+                guardiansSignaturesReached;
+              const executeLegalVetoExtensionEnabled =
+                !proposal.extendedLegalVeto && proposal.status === PROPOSAL_STATES.LegalVetoPeriod;
+              const executeProposalEnabled =
+                proposal.status === PROPOSAL_STATES.Ready &&
+                (isAddressEqual(proposal.executor as Hex, user.address as Hex) ||
+                  isAddressEqual(proposal.executor as Hex, zeroAddress));
 
-            return (
-              <>
-                <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
-                  <Card className="pb-10">
-                    <CardHeader>
-                      <CardTitle>Proposal Details</CardTitle>
-                    </CardHeader>
-                    <CardContent>
-                      <div className="space-y-6">
-                        <div className="flex justify-between">
-                          <span>Current Version:</span>
-                          <span className="w-1/2 break-words text-right">
-                            {proposal.currentVersion}
-                          </span>
-                        </div>
-                        <div className="flex justify-between">
-                          <span>Proposed Version:</span>
-                          <span className="w-1/2 break-words text-right">
-                            {proposal.proposedVersion}
-                          </span>
-                        </div>
-                        <div className="flex justify-between">
-                          <span>Proposal ID:</span>
-                          <span className="w-1/2 justify-end break-words">{proposalId}</span>
-                        </div>
-                        <div className="flex justify-between">
-                          <span>Proposed On:</span>
-                          <div className="flex w-1/2 flex-col break-words text-right">
-                            <span>{new Date(proposal.proposedOn).toISOString()}</span>
-                            <span>({dateToUnixTimestamp(new Date(proposal.proposedOn))})</span>
+              return (
+                <>
+                  <div className="grid grid-cols-1 gap-4 md:grid-cols-2">
+                    <Card className="pb-10">
+                      <CardHeader>
+                        <CardTitle>Proposal Details</CardTitle>
+                      </CardHeader>
+                      <CardContent>
+                        <div className="space-y-6">
+                          <div className="flex justify-between">
+                            <span>Current Version:</span>
+                            <span className="w-1/2 break-words text-right">
+                              {proposal.currentVersion}
+                            </span>
+                          </div>
+                          <div className="flex justify-between">
+                            <span>Proposed Version:</span>
+                            <span className="w-1/2 break-words text-right">
+                              {proposal.proposedVersion}
+                            </span>
+                          </div>
+                          <div className="flex justify-between">
+                            <span>Proposal ID:</span>
+                            <span className="w-1/2 justify-end break-words">{proposalId}</span>
+                          </div>
+                          <div className="flex justify-between">
+                            <span>Proposed On:</span>
+                            <div className="flex w-1/2 flex-col break-words text-right">
+                              <span>{new Date(proposal.proposedOn).toISOString()}</span>
+                              <span>({dateToUnixTimestamp(new Date(proposal.proposedOn))})</span>
+                            </div>
+                          </div>
+                          <div className="flex justify-between">
+                            <span>Executor:</span>
+                            <span className="w-1/2 break-words text-right">
+                              {isAddressEqual(proposal.executor as `0x${string}`, zeroAddress)
+                                ? "Anyone"
+                                : proposal.executor}
+                            </span>
+                          </div>
+                          <div className="flex justify-between">
+                            <span>Transaction hash:</span>
+                            <TxLink txid={proposal.transactionHash} network={ethNetwork} />
                           </div>
                         </div>
-                        <div className="flex justify-between">
-                          <span>Executor:</span>
-                          <span className="w-1/2 break-words text-right">
-                            {isAddressEqual(proposal.executor as `0x${string}`, zeroAddress)
-                              ? "Anyone"
-                              : proposal.executor}
-                          </span>
+                      </CardContent>
+                    </Card>
+                    <Card className="pb-10">
+                      <CardHeader className="pt-7">
+                        <ProposalState status={proposal.status} times={proposal.statusTimes} />
+                        <CardTitle>Proposal Status</CardTitle>
+                      </CardHeader>
+                      <CardContent>
+                        <div className="space-y-5">
+                          <VotingStatusIndicator
+                            label="Security Council Approvals"
+                            signatures={proposal.signatures.approveUpgradeSecurityCouncil.length}
+                            necessarySignatures={NECESSARY_SECURITY_COUNCIL_SIGNATURES}
+                          />
+                          <VotingStatusIndicator
+                            label="Guardian Approvals"
+                            signatures={proposal.signatures.approveUpgradeGuardians.length}
+                            necessarySignatures={NECESSARY_GUARDIAN_SIGNATURES}
+                          />
+                          <VotingStatusIndicator
+                            label="Extend Legal Veto Approvals"
+                            signatures={proposal.signatures.extendLegalVetoPeriod.length}
+                            necessarySignatures={NECESSARY_LEGAL_VETO_SIGNATURES}
+                          />
                         </div>
-                        <div className="flex justify-between">
-                          <span>Transaction hash:</span>
-                          <TxLink txid={proposal.transactionHash} network={ethNetwork} />
-                        </div>
-                      </div>
-                    </CardContent>
-                  </Card>
-                  <Card className="pb-10">
-                    <CardHeader className="pt-7">
-                      <ProposalState status={proposal.status} times={proposal.statusTimes} />
-                      <CardTitle>Proposal Status</CardTitle>
-                    </CardHeader>
-                    <CardContent>
-                      <div className="space-y-5">
-                        <VotingStatusIndicator
-                          label="Security Council Approvals"
-                          signatures={proposal.signatures.approveUpgradeSecurityCouncil.length}
-                          necessarySignatures={NECESSARY_SECURITY_COUNCIL_SIGNATURES}
-                        />
-                        <VotingStatusIndicator
-                          label="Guardian Approvals"
-                          signatures={proposal.signatures.approveUpgradeGuardians.length}
-                          necessarySignatures={NECESSARY_GUARDIAN_SIGNATURES}
-                        />
-                        <VotingStatusIndicator
-                          label="Extend Legal Veto Approvals"
-                          signatures={proposal.signatures.extendLegalVetoPeriod.length}
-                          necessarySignatures={NECESSARY_LEGAL_VETO_SIGNATURES}
-                        />
-                      </div>
-                    </CardContent>
-                  </Card>
-                  <Card className="pb-10">
-                    <CardHeader>
-                      <CardTitle>
-                        {user.role === "visitor" && "No role actions"}
-                        {user.role === "guardian" && "Guardian Actions"}
-                        {user.role === "securityCouncil" && "Security Council Actions"}
-                      </CardTitle>
-                    </CardHeader>
-                    <CardContent className="flex flex-col space-y-3">
-                      {user.role === "guardian" && (
-                        <SignButton
+                      </CardContent>
+                    </Card>
+                    <Card className="pb-10">
+                      <CardHeader>
+                        <CardTitle>
+                          {user.role === "visitor" && "No role actions"}
+                          {user.role === "guardian" && "Guardian Actions"}
+                          {user.role === "securityCouncil" && "Security Council Actions"}
+                        </CardTitle>
+                      </CardHeader>
+                      <CardContent className="flex flex-col space-y-3">
+                        {user.role === "guardian" && (
+                          <SignButton
+                            proposalId={proposalId}
+                            contractData={{
+                              actionName: "ExtendLegalVetoPeriod",
+                              address: addresses.guardians,
+                              name: "Guardians",
+                            }}
+                            disabled={!signLegalVetoEnabled}
+                            postAction={$path("/app/proposals/:id", { id: proposalId })}
+                          >
+                            Approve extend veto period
+                          </SignButton>
+                        )}
+                        {user.role === "guardian" && (
+                          <SignButton
+                            proposalId={proposalId}
+                            contractData={{
+                              actionName: "ApproveUpgradeGuardians",
+                              address: addresses.guardians,
+                              name: "Guardians",
+                            }}
+                            disabled={!signProposalEnabled}
+                            postAction={$path("/app/proposals/:id", { id: proposalId })}
+                          >
+                            Approve proposal
+                          </SignButton>
+                        )}
+                        {user.role === "securityCouncil" && (
+                          <SignButton
+                            proposalId={proposalId}
+                            contractData={{
+                              actionName: "ApproveUpgradeSecurityCouncil",
+                              address: addresses.council,
+                              name: "SecurityCouncil",
+                            }}
+                            disabled={!signProposalEnabled}
+                            postAction={$path("/app/proposals/:id", { id: proposalId })}
+                          >
+                            Approve proposal
+                          </SignButton>
+                        )}
+                      </CardContent>
+                    </Card>
+                    <Card className="pb-10">
+                      <CardHeader>
+                        <CardTitle>Proposal Actions</CardTitle>
+                      </CardHeader>
+                      <CardContent className="flex flex-col space-y-3">
+                        <ContractWriteButton
+                          target={addresses.council}
+                          signatures={proposal.signatures.approveUpgradeSecurityCouncil}
                           proposalId={proposalId}
-                          contractData={{
-                            actionName: "ExtendLegalVetoPeriod",
-                            address: addresses.guardians,
-                            name: "Guardians",
-                          }}
-                          disabled={!signLegalVetoEnabled}
-                          postAction={$path("/app/proposals/:id", { id: proposalId })}
+                          functionName={"approveUpgradeSecurityCouncil"}
+                          abiName="council"
+                          threshold={NECESSARY_SECURITY_COUNCIL_SIGNATURES}
+                          disabled={!executeSecurityCouncilApprovalEnabled}
                         >
-                          Approve extend veto period
-                        </SignButton>
-                      )}
-                      {user.role === "guardian" && (
-                        <SignButton
+                          Execute security council approval
+                        </ContractWriteButton>
+
+                        <ContractWriteButton
+                          target={addresses.guardians}
+                          signatures={proposal.signatures.approveUpgradeGuardians}
                           proposalId={proposalId}
-                          contractData={{
-                            actionName: "ApproveUpgradeGuardians",
-                            address: addresses.guardians,
-                            name: "Guardians",
-                          }}
-                          disabled={!signProposalEnabled}
-                          postAction={$path("/app/proposals/:id", { id: proposalId })}
+                          functionName={"approveUpgradeGuardians"}
+                          abiName="guardians"
+                          threshold={NECESSARY_GUARDIAN_SIGNATURES}
+                          disabled={!executeGuardiansApprovalEnabled}
                         >
-                          Approve proposal
-                        </SignButton>
-                      )}
-                      {user.role === "securityCouncil" && (
-                        <SignButton
+                          Execute guardian approval
+                        </ContractWriteButton>
+
+                        <ContractWriteButton
+                          target={addresses.guardians}
+                          signatures={proposal.signatures.extendLegalVetoPeriod}
                           proposalId={proposalId}
-                          contractData={{
-                            actionName: "ApproveUpgradeSecurityCouncil",
-                            address: addresses.council,
-                            name: "SecurityCouncil",
-                          }}
-                          disabled={!signProposalEnabled}
-                          postAction={$path("/app/proposals/:id", { id: proposalId })}
+                          functionName={"extendLegalVeto"}
+                          abiName="guardians"
+                          threshold={NECESSARY_LEGAL_VETO_SIGNATURES}
+                          disabled={!executeLegalVetoExtensionEnabled}
                         >
-                          Approve proposal
-                        </SignButton>
-                      )}
-                    </CardContent>
-                  </Card>
-                  <Card className="pb-10">
-                    <CardHeader>
-                      <CardTitle>Proposal Actions</CardTitle>
-                    </CardHeader>
-                    <CardContent className="flex flex-col space-y-3">
-                      <ContractWriteButton
-                        target={addresses.council}
-                        signatures={proposal.signatures.approveUpgradeSecurityCouncil}
-                        proposalId={proposalId}
-                        functionName={"approveUpgradeSecurityCouncil"}
-                        abiName="council"
-                        threshold={NECESSARY_SECURITY_COUNCIL_SIGNATURES}
-                        disabled={!executeSecurityCouncilApprovalEnabled}
-                      >
-                        Execute security council approval
-                      </ContractWriteButton>
+                          Execute legal veto extension
+                        </ContractWriteButton>
+                        <ExecuteUpgradeButton
+                          target={addresses.upgradeHandler}
+                          proposalCalldata={proposal.raw}
+                          disabled={!executeProposalEnabled}
+                        >
+                          Execute upgrade
+                        </ExecuteUpgradeButton>
+                      </CardContent>
+                    </Card>
+                  </div>
+                  <div className="pt-4">
+                    <h2 className="font-bold text-3xl">Upgrade Analysis</h2>
+                    <Tabs className="mt-4 flex" defaultValue="facet-changes">
+                      <TabsList className="mt-12 mr-6">
+                        <TabsTrigger value="facet-changes">Facet Changes</TabsTrigger>
+                        <TabsTrigger value="system-contract-changes">
+                          System Contract Changes
+                        </TabsTrigger>
+                        <TabsTrigger value="field-changes">Field Changes</TabsTrigger>
+                        <TabsTrigger value="field-storage-changes">
+                          Field Storage Changes
+                        </TabsTrigger>
+                        <TabsTrigger value="raw-data">Raw Data</TabsTrigger>
+                      </TabsList>
+                      <TabsContent value="facet-changes" className="w-full">
+                        <Card className="pb-8">
+                          <CardHeader>
+                            <CardTitle>Facet Changes</CardTitle>
+                          </CardHeader>
+                          <CardContent className="pt-4">
+                            <FacetChangesTable data={reports.facetChanges} />
+                          </CardContent>
+                        </Card>
+                      </TabsContent>
+                      <TabsContent value="system-contract-changes" className="w-full">
+                        <Card className="pb-8">
+                          <CardHeader>
+                            <CardTitle>System Contract Changes</CardTitle>
+                          </CardHeader>
+                          <CardContent className="pt-4">
+                            <SystemContractChangesTable data={reports.systemContractChanges} />
+                          </CardContent>
+                        </Card>
+                      </TabsContent>
+                      <TabsContent value="field-changes" className="w-full">
+                        <Card className="pb-8">
+                          <CardHeader>
+                            <CardTitle>Field Changes</CardTitle>
+                          </CardHeader>
+                          <CardContent className="pt-4">
+                            <FieldChangesTable data={reports.fieldChanges} />
+                          </CardContent>
+                        </Card>
+                      </TabsContent>
+                      <TabsContent value="field-storage-changes" className="w-full">
+                        <Card className="pb-8">
+                          <CardHeader>
+                            <CardTitle>Field Storage Changes</CardTitle>
+                          </CardHeader>
+                          <CardContent className="pt-4">
+                            <FieldStorageChangesTable data={reports.fieldStorageChanges} />{" "}
+                          </CardContent>
+                        </Card>
+                      </TabsContent>
 
-                      <ContractWriteButton
-                        target={addresses.guardians}
-                        signatures={proposal.signatures.approveUpgradeGuardians}
-                        proposalId={proposalId}
-                        functionName={"approveUpgradeGuardians"}
-                        abiName="guardians"
-                        threshold={NECESSARY_GUARDIAN_SIGNATURES}
-                        disabled={!executeGuardiansApprovalEnabled}
-                      >
-                        Execute guardian approval
-                      </ContractWriteButton>
-
-                      <ContractWriteButton
-                        target={addresses.guardians}
-                        signatures={proposal.signatures.extendLegalVetoPeriod}
-                        proposalId={proposalId}
-                        functionName={"extendLegalVeto"}
-                        abiName="guardians"
-                        threshold={NECESSARY_LEGAL_VETO_SIGNATURES}
-                        disabled={!executeLegalVetoExtensionEnabled}
-                      >
-                        Execute legal veto extension
-                      </ContractWriteButton>
-                      <ExecuteUpgradeButton
-                        target={addresses.upgradeHandler}
-                        proposalCalldata={proposal.raw}
-                        disabled={!executeProposalEnabled}
-                      >
-                        Execute upgrade
-                      </ExecuteUpgradeButton>
-                    </CardContent>
-                  </Card>
-                </div>
-                <div className="pt-4">
-                  <h2 className="font-bold text-3xl">Upgrade Analysis</h2>
-                  <Tabs className="mt-4 flex" defaultValue="facet-changes">
-                    <TabsList className="mt-12 mr-6">
-                      <TabsTrigger value="facet-changes">Facet Changes</TabsTrigger>
-                      <TabsTrigger value="system-contract-changes">
-                        System Contract Changes
-                      </TabsTrigger>
-                      <TabsTrigger value="field-changes">Field Changes</TabsTrigger>
-                      <TabsTrigger value="field-storage-changes">Field Storage Changes</TabsTrigger>
-                      <TabsTrigger value="raw-data">Raw Data</TabsTrigger>
-                    </TabsList>
-                    <TabsContent value="facet-changes" className="w-full">
-                      <Card className="pb-8">
-                        <CardHeader>
-                          <CardTitle>Facet Changes</CardTitle>
-                        </CardHeader>
-                        <CardContent className="pt-4">
-                          <FacetChangesTable data={reports.facetChanges} />
-                        </CardContent>
-                      </Card>
-                    </TabsContent>
-                    <TabsContent value="system-contract-changes" className="w-full">
-                      <Card className="pb-8">
-                        <CardHeader>
-                          <CardTitle>System Contract Changes</CardTitle>
-                        </CardHeader>
-                        <CardContent className="pt-4">
-                          <SystemContractChangesTable data={reports.systemContractChanges} />
-                        </CardContent>
-                      </Card>
-                    </TabsContent>
-                    <TabsContent value="field-changes" className="w-full">
-                      <Card className="pb-8">
-                        <CardHeader>
-                          <CardTitle>Field Changes</CardTitle>
-                        </CardHeader>
-                        <CardContent className="pt-4">
-                          <FieldChangesTable data={reports.fieldChanges} />
-                        </CardContent>
-                      </Card>
-                    </TabsContent>
-                    <TabsContent value="field-storage-changes" className="w-full">
-                      <Card className="pb-8">
-                        <CardHeader>
-                          <CardTitle>Field Storage Changes</CardTitle>
-                        </CardHeader>
-                        <CardContent className="pt-4">
-                          <FieldStorageChangesTable data={reports.fieldStorageChanges} />{" "}
-                        </CardContent>
-                      </Card>
-                    </TabsContent>
-
-                    <TabsContent value="raw-data" className="w-full">
-                      <Card className="pb-8">
-                        <CardHeader>
-                          <CardTitle>Raw Data</CardTitle>
-                        </CardHeader>
-                        <CardContent className="pt-4">
-                          <RawStandardUpgrade encoded={proposal.raw} />
-                        </CardContent>
-                      </Card>
-                    </TabsContent>
-                  </Tabs>
-                </div>
-              </>
-            );
-          }}
-        </Await>
-      </Suspense>
+                      <TabsContent value="raw-data" className="w-full">
+                        <Card className="pb-8">
+                          <CardHeader>
+                            <CardTitle>Raw Data</CardTitle>
+                          </CardHeader>
+                          <CardContent className="pt-4">
+                            <RawStandardUpgrade encoded={proposal.raw} />
+                          </CardContent>
+                        </Card>
+                      </TabsContent>
+                    </Tabs>
+                  </div>
+                </>
+              );
+            }}
+          </Await>
+        </Suspense>
+      </div>
     </div>
   );
 }
